+++ packages/common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mlightcad/common
 
+## 1.1.4
+
+### Patch Changes
+
+- fix: fix issue that AcDbBatchProcessing.processChunk doesn't wait for all chunks processed and then return
+
 ## 1.1.3
 
 ### Patch Changes