+++ packages/graphic-interface/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mlightcad/graphic-interface
 
+## 2.0.5
+
+### Patch Changes
+
+- @mlightcad/geometry-engine@2.0.5
+
 ## 2.0.4
 
 ### Patch Changes